Add a generateHyperLogLog function
Unlike the `HyperLogLog` data constructor or its `mempty` implementation, which
allow constructing `HyperLogLog` values using fixed `SipKey`s,
`generateHyperLogLog` builds a `HyperLogLog` value using a randomly generated
`SipKey` that is derived from system entropy (using the `entropy` library).
This prevents an adversary from exploiting the potential for exponentially
inaccurate results if a fixed `SipKey` is used, as described in
https://eprint.iacr.org/2021/1139.

I have not gone as far as deprecating the parts of the API that allow using
fixed `SipKey`s, as there are legitimate use cases for doing so (e.g., data
collection across processes). In lieu of this, I have added prominent
disclaimers to the relevant Haddocks.

Fixes #32.